PDBApi-master/PDBApi-master/requirements.txt
certifi==2019.9.11
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djongo==1.2.36
idna==2.8
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.2
pkg-resources==0.0.0
PyJWT==1.7.1
pymongo==3.9.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.25.6
